extract reusable Input component and built TextInput on top of it
