+++ applications/launchpad_v2/src/containers/WalletContainer/PasswordBox.tsx
@@ -1,6 +1,7 @@
 import { useTheme } from 'styled-components'
-import { useState, ChangeEvent, SyntheticEvent } from 'react'
+import { useState, SyntheticEvent } from 'react'
 
+import TextInput from '../../components/Inputs/TextInput'
 import Box from '../../components/Box'
 import Text from '../../components/Text'
 import Button from '../../components/Button'
@@ -19,9 +20,7 @@ const PasswordBox = ({
 }) => {
   const theme = useTheme()
   const [password, setPassword] = useState('')
-  const updatePassword = (event: ChangeEvent<HTMLInputElement>) => {
-    const v = event.target.value
-
+  const updatePassword = (v: string) => {
     setPassword(v)
   }
 
@@ -41,15 +40,13 @@ const PasswordBox = ({
       </Text>
       <Text>{t.wallet.password.cta}</Text>
       <form onSubmit={formSubmitHandler}>
-        <input
+        <TextInput
           type='password'
           onChange={updatePassword}
           value={password}
           disabled={pending}
           placeholder={t.wallet.password.placeholderCta}
           style={{
-            marginTop: theme.spacingVertical(3),
-            marginBottom: theme.spacingVertical(1.5),
             width: '100%',
           }}
         />
